Update 3/30/25
Updated the card renderer so that it can now add card art. (Note: This means that every single card with placeholder art has changed by a few pixels!)
Added an Injury mechanic to Hana's cards.
Nerfed Brainwashed.
Prototype Pseudo-Humanoid Assault Drone's cannon card now requires it to be controlled.
Agriajuda Terraformed Region is now a transient that can't be included in a deck.